Clean up lint errors, reorganize components
