--- client/src/pages/Profile.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Profile = (props) => {
-  return (
-    <div>
-      <h2>Profile Page</h2>
-    </div>
-  );
-};
-
-export default Profile;
